App constants and config changes
