+++ src/config/chartConfig.tsx
@@ -9,7 +9,7 @@ import {
 } from "chart.js";
 import ChartDataLabels from "chartjs-plugin-datalabels";
 import {
-  SESSIONS_CHART,
+  SESSIONS_CHART_DEFAULT,
   TOTAL_ORDERS_PER_MINUTE,
 } from "../constants/appConstants";
 import { externalTooltipHandler } from "../components/utils/Utils";
@@ -24,7 +24,10 @@ Chart.register(
   ChartDataLabels,
 );
 
-export const BAR_CHART_OPTIONS: Chart.ChartOptions = {
+export const BAR_CHART_OPTIONS: Chart.ChartOptions = (
+  showDataLabels = false,
+) => ({
+  // indexAxis: 'y',
   responsive: true,
   maintainAspectRatio: false,
   plugins: {
@@ -43,7 +46,7 @@ export const BAR_CHART_OPTIONS: Chart.ChartOptions = {
               text: dataset.label,
               fillStyle: "#30343B",
               fontColor: dataset.backgroundColor,
-              lineWidth: SESSIONS_CHART.LEGEND_LINE_WIDTH,
+              lineWidth: SESSIONS_CHART_DEFAULT.LEGEND_LINE_WIDTH,
               strokeStyle: dataset.backgroundColor,
             }));
           }
@@ -51,7 +54,7 @@ export const BAR_CHART_OPTIONS: Chart.ChartOptions = {
       },
     },
     datalabels: {
-      display: false,
+      display: showDataLabels,
       formatter: (_, context) =>
         context.chart.data.dataset?.[0]?.data?.[context.dataIndex],
       color: "black",
@@ -72,6 +75,7 @@ export const BAR_CHART_OPTIONS: Chart.ChartOptions = {
       },
     },
     y: {
+      beginAtZero: true,
       grid: {
         display: true,
       },
@@ -79,52 +83,300 @@ export const BAR_CHART_OPTIONS: Chart.ChartOptions = {
         display: false,
       },
       ticks: {
-        display: false,
-        stepSize: SESSIONS_CHART.STEP_SIZE,
-        count: SESSIONS_CHART.TICK_COUNT,
+        display: true,
+        stepSize: SESSIONS_CHART_DEFAULT.STEP_SIZE,
+        count: SESSIONS_CHART_DEFAULT.TICK_COUNT,
+        callback: function (value) {
+          if (value >= 1000) {
+            return value / 1000 + "K";
+          } else {
+            return value;
+          }
+        },
       },
     },
   },
   datasets: {
     bar: {
-      barPercentage: SESSIONS_CHART.BAR_PERCENT,
-      maxBarThickness: SESSIONS_CHART.MAX_BAR_THICKNESS,
-      borderRadius: SESSIONS_CHART.BAR_BORDER_RADIUS,
-      categoryPercentage: SESSIONS_CHART.CATEGORY_PERCENT,
+      barPercentage: SESSIONS_CHART_DEFAULT.BAR_PERCENT,
+      maxBarThickness: SESSIONS_CHART_DEFAULT.MAX_BAR_THICKNESS,
+      borderRadius: SESSIONS_CHART_DEFAULT.BAR_BORDER_RADIUS,
+      categoryPercentage: SESSIONS_CHART_DEFAULT.CATEGORY_PERCENT,
     },
   },
+});
+
+//OPM Charts
+export const OPM_BAR_CHART_OPTIONS: Chart.ChartOptions = (
+  isMobile: boolean,
+  showDataLabels = false,
+) => {
+  const existingBarChartOptions = {
+    ...BAR_CHART_OPTIONS(showDataLabels && !isMobile),
+  };
+  return {
+    ...existingBarChartOptions,
+    plugins: {
+      ...existingBarChartOptions.plugins,
+      legend: {
+        display: false,
+      },
+      tooltip: {
+        enabled: false,
+        external: (_) => externalTooltipHandler(_, "opm_bar"),
+      },
+      datalabels: {
+        display: showDataLabels,
+        formatter: (_, context) =>
+          context.chart.data.dataset?.[0]?.data?.[context.dataIndex],
+        align: "top",
+        anchor: "center",
+        font: {
+          size: "12",
+          color: "#000000",
+        },
+      },
+      title: {
+        display: true,
+        color: "#FAF9F6",
+        text: TOTAL_ORDERS_PER_MINUTE,
+        position: "bottom",
+        font: {
+          size: "12",
+        },
+        padding: 20,
+      },
+    },
+    scales: {
+      ...existingBarChartOptions.scales,
+      y: {
+        grid: {
+          color: "#00000033",
+        },
+        border: {
+          display: false,
+        },
+      },
+    },
+  };
 };
 
-export const OPM_BAR_CHART_OPTIONS: Chart.ChartOptions = {
-  ...BAR_CHART_OPTIONS,
+export const OPM_OPTIONS = (isMobile: boolean, showDataLabels = false) => ({
+  responsive: true,
+  maintainAspectRatio: false,
+  layout: {
+    padding: isMobile
+      ? {
+          left: 20,
+          right: 20,
+          top: 10,
+          bottom: 40,
+        }
+      : {
+          left: 30,
+          right: 50,
+          top: 50,
+          bottom: 20,
+        },
+  },
+  scales: {
+    y: {
+      grid: {
+        color: "#00000033",
+      },
+      border: {
+        display: false,
+      },
+    },
+    x: {
+      grid: {
+        display: false,
+      },
+      title: {
+        display: true,
+        color: "#FAF9F6",
+        text: TOTAL_ORDERS_PER_MINUTE,
+        padding: isMobile ? { top: 35, bottom: 35 } : { top: 35 },
+      },
+    },
+  },
   plugins: {
-    ...BAR_CHART_OPTIONS.plugins,
     legend: {
       display: false,
     },
     tooltip: {
       enabled: false,
-      external: (_) => externalTooltipHandler(_, "opm_bar"),
+      external: (_) => externalTooltipHandler(_, "opm"),
     },
-    title: {
-      display: true,
-      color: "#FAF9F6",
-      text: TOTAL_ORDERS_PER_MINUTE,
-      position: "bottom",
+    datalabels: {
+      display: showDataLabels,
+      formatter: (_, context) =>
+        context.chart.data.dataset?.[0]?.data?.[context.dataIndex],
+      align: "top",
+      anchor: "center",
       font: {
         size: "12",
       },
-      padding: 20,
     },
   },
+  elements: {
+    point: {
+      radius: 4,
+      backgroundColor: "white",
+    },
+  },
+});
+
+export const OPM_COMPARISON_OPTIONS = ({
+  apiResponse,
+  startDate,
+  endDate,
+  isMobile,
+  showDataLabels = false,
+}) => ({
+  responsive: true,
+  maintainAspectRatio: false,
+  layout: {
+    padding: isMobile
+      ? {
+          left: 20,
+          right: 20,
+          top: 10,
+          bottom: 40,
+        }
+      : {
+          left: 30,
+          right: 50,
+          top: 50,
+          bottom: 20,
+        },
+  },
   scales: {
-    ...BAR_CHART_OPTIONS.scales,
+    x: {
+      grid: {
+        display: false,
+      },
+      title: {
+        display: true,
+        text: TOTAL_ORDERS_PER_MINUTE,
+        color: "#E8E8E8",
+        position: "left",
+        padding: isMobile ? { top: 20, bottom: 20 } : { top: 50 },
+      },
+    },
     y: {
-      ...BAR_CHART_OPTIONS.scales.y,
-      ticks: {
-        ...BAR_CHART_OPTIONS.scales.y.ticks,
-        stepSize: 100,
+      border: {
+        display: false,
       },
     },
   },
+  elements: {
+    point: {
+      radius: 4,
+      backgroundColor: "white",
+    },
+  },
+  plugins: {
+    datalabels: {
+      display: showDataLabels,
+      formatter: (_, context) =>
+        context.chart.data.dataset?.[0]?.data?.[context.dataIndex],
+      align: "top",
+      anchor: "center",
+      font: {
+        size: "12",
+      },
+    },
+    legend: {
+      display: true,
+      position: isMobile ? "top" : "bottom",
+      align: "start",
+      labels: {
+        boxWidth: 30,
+        backgroundColor: "transparent",
+        generateLabels: () => {
+          return Object.keys(apiResponse).map((_, index) => ({
+            text:
+              index === 0
+                ? startDate?.toLocaleString("en-US", {
+                    year: "numeric",
+                    month: "2-digit",
+                    day: "2-digit",
+                  })
+                : endDate?.toLocaleString("en-US", {
+                    year: "numeric",
+                    month: "2-digit",
+                    day: "2-digit",
+                  }),
+            fillStyle: "transparent",
+            lineWidth: 2,
+            fontColor: index === 0 ? "#6370FF" : "#FDA44F",
+            strokeStyle: index === 0 ? "#6370FF" : "#FDA44F",
+          }));
+        },
+      },
+    },
+    tooltip: {
+      enabled: false,
+      external: (_) => externalTooltipHandler(_, "opmComparison"),
+    },
+  },
+});
+
+export const OPM_COMPARISON_OPTIONS_HOME = ({
+  apiResponse,
+  startDate,
+  endDate,
+  isMobile,
+  showDataLabels = false,
+}) => {
+  const options = OPM_COMPARISON_OPTIONS({
+    apiResponse,
+    startDate,
+    endDate,
+    isMobile,
+  });
+  return {
+    ...options,
+    layout: {
+      ...options.layout,
+      padding: isMobile
+        ? { left: 10, right: 20, top: 15, bottom: 0 }
+        : {
+            left: 30,
+            right: 50,
+            top: 35,
+            bottom: 0,
+          },
+    },
+    scales: {
+      ...options.scales,
+      x: {
+        ...options.scales.x,
+        title: {
+          ...options.scales.x.title,
+          padding: isMobile
+            ? { top: 20, bottom: 20 }
+            : { left: 50, top: 35, bottom: -23 },
+        },
+      },
+    },
+    plugins: {
+      ...options.plugins,
+      datalabels: {
+        display: showDataLabels,
+        formatter: (_, context: any) =>
+          context.chart.data.dataset?.[0]?.data?.[context.dataIndex],
+        align: "top",
+        anchor: "center",
+        font: {
+          size: "10",
+        },
+      },
+      legend: {
+        ...options.plugins.legend,
+        position: "bottom",
+      },
+    },
+  };
 };
