+++ src/constants/appConstants.tsx
@@ -1,5 +1,3 @@
-import { externalTooltipHandler } from "../components/utils/Utils";
-
 export const SESSIONS: string = "Sessions";
 export const DURATION: string = "Duration";
 export const TIME: string = "Time";
@@ -18,7 +16,7 @@ export const TOTAL_SESSIONS_PER_MIN_PRIMARY: string =
   "Azure Primary (Sessions/Min)";
 export const TOTAL_SESSIONS_PER_MIN_SECONDARY: string =
   "Azure Secondary (Sessions/Min)";
-export const TOTAL_SESSIONS_PER_MINUTE: string = "Total Session per minute";
+export const TOTAL_SESSIONS_PER_MINUTE: string = "Total Sessions per minute";
 export const TOTAL_ORDERS_PER_MINUTE: string = "Total Orders Per Minute";
 export const MINS = "Mins";
 export const LASTDAY: string = "Last Day";
@@ -111,14 +109,15 @@ export const CHART = {
   ROTATION_0: 0,
   ROTATION_270: 270,
 };
-export const SESSIONS_CHART = {
-  BAR_PERCENT: 0.9,
-  MAX_BAR_THICKNESS: 40,
+
+export const SESSIONS_CHART_DEFAULT = {
   BAR_BORDER_RADIUS: 6,
   STEP_SIZE: 5000,
   TICK_COUNT: 5,
-  CATEGORY_PERCENT: 0.6,
   LEGEND_LINE_WIDTH: 2,
+  BAR_PERCENT: 0.95,
+  MAX_BAR_THICKNESS: 50,
+  CATEGORY_PERCENT: 0.9,
 };
 
 export const DURATIONS = {
@@ -187,225 +186,6 @@ export const ORDER_STATUS_LIST = [
     Description: "Shipped and invoiced",
   },
 ];
-export const OPM_OPTIONS = (isMobile: boolean, showDataLabels = false) => ({
-  responsive: true,
-  maintainAspectRatio: false,
-  layout: {
-    padding: isMobile
-      ? {
-        left: 20,
-        right: 20,
-        top: 10,
-        bottom: 40,
-      }
-      : {
-        left: 30,
-        right: 50,
-        top: 50,
-        bottom: 20,
-      },
-  },
-  scales: {
-    y: {
-      grid: {
-        color: "#00000033",
-      },
-      border: {
-        display: false,
-      },
-    },
-    x: {
-      grid: {
-        display: false,
-      },
-      title: {
-        display: true,
-        color: "#FAF9F6",
-        text: "Total Orders Per Minute",
-        padding: isMobile ? { top: 35, bottom: 35 } : { top: 35 },
-      },
-    },
-  },
-  plugins: {
-    legend: {
-      display: false,
-    },
-    tooltip: {
-      enabled: false,
-      external: (_) => externalTooltipHandler(_, "opm"),
-    },
-    datalabels: {
-      display: showDataLabels,
-      formatter: (_, context) =>
-        context.chart.data.dataset?.[0]?.data?.[context.dataIndex],
-      align: "top",
-      anchor: "center",
-      font: {
-        size: "12",
-      },
-    },
-  },
-  elements: {
-    point: {
-      radius: 4,
-      backgroundColor: "white",
-    },
-  },
-});
-
-export const OPM_COMPARISON_OPTIONS = ({
-  apiResponse,
-  startDate,
-  endDate,
-  isMobile,
-  showDataLabels = false,
-}) => ({
-  responsive: true,
-  maintainAspectRatio: false,
-  layout: {
-    padding: isMobile
-      ? {
-        left: 20,
-        right: 20,
-        top: 10,
-        bottom: 40,
-      }
-      : {
-        left: 30,
-        right: 50,
-        top: 50,
-        bottom: 20,
-      },
-  },
-  scales: {
-    x: {
-      grid: {
-        display: false,
-      },
-      title: {
-        display: true,
-        text: "Total Orders Per Minute",
-        color: "#E8E8E8",
-        position: "left",
-        padding: isMobile ? { top: 20, bottom: 20 } : { top: 50 },
-      },
-    },
-    y: {
-      border: {
-        display: false,
-      },
-    },
-  },
-  elements: {
-    point: {
-      radius: 4,
-      backgroundColor: "white",
-    },
-  },
-  plugins: {
-    datalabels: {
-      display: showDataLabels,
-      formatter: (_, context) =>
-        context.chart.data.dataset?.[0]?.data?.[context.dataIndex],
-      align: "top",
-      anchor: "center",
-      font: {
-        size: "12",
-      },
-    },
-    legend: {
-      display: true,
-      position: isMobile ? "top" : "bottom",
-      align: "start",
-      labels: {
-        boxWidth: 30,
-        backgroundColor: "transparent",
-        generateLabels: () => {
-          return Object.keys(apiResponse).map((_, index) => ({
-            text:
-              index === 0
-                ? startDate?.toLocaleString("en-US", {
-                  year: "numeric",
-                  month: "2-digit",
-                  day: "2-digit",
-                })
-                : endDate?.toLocaleString("en-US", {
-                  year: "numeric",
-                  month: "2-digit",
-                  day: "2-digit",
-                }),
-            fillStyle: "transparent",
-            lineWidth: 2,
-            fontColor: index === 0 ? "#6370FF" : "#FDA44F",
-            strokeStyle: index === 0 ? "#6370FF" : "#FDA44F",
-          }));
-        },
-      },
-    },
-    tooltip: {
-      enabled: false,
-      external: (_) => externalTooltipHandler(_, "opmComparison"),
-    },
-  },
-});
-
-export const OPM_COMPARISON_OPTIONS_HOME = ({
-  apiResponse,
-  startDate,
-  endDate,
-  isMobile,
-  showDataLabels = false,
-}) => {
-  const options = OPM_COMPARISON_OPTIONS({
-    apiResponse,
-    startDate,
-    endDate,
-    isMobile,
-  });
-  return {
-    ...options,
-    layout: {
-      ...options.layout,
-      padding: isMobile
-        ? { left: 10, right: 20, top: 15, bottom: 0 }
-        : {
-          left: 30,
-          right: 50,
-          top: 35,
-          bottom: 0,
-        },
-    },
-    scales: {
-      ...options.scales,
-      x: {
-        ...options.scales.x,
-        title: {
-          ...options.scales.x.title,
-          padding: isMobile
-            ? { top: 20, bottom: 20 }
-            : { left: 50, top: 35, bottom: -23 },
-        },
-      },
-    },
-    plugins: {
-      ...options.plugins,
-      datalabels: {
-        display: showDataLabels,
-        formatter: (_, context: any) =>
-          context.chart.data.dataset?.[0]?.data?.[context.dataIndex],
-        align: "top",
-        anchor: "center",
-        font: {
-          size: "10",
-        },
-      },
-      legend: {
-        ...options.plugins.legend,
-        position: "bottom",
-      },
-    },
-  };
-};
 
 export const ORDER_STATUS: string = "Order Status";
 export const SESSIONS_TABS = [
@@ -414,44 +194,49 @@ export const SESSIONS_TABS = [
   { header: BOTH },
 ];
 
-export const QUICK_LINKS = [{
-  id: 0,
-  header: "Alerting & Monitoring Tools",
-  links: [
-    { name: "Azure Splunk", link: import.meta.env.VITE_AZURE_SPLUNK },
-    { name: "Azure 2 Splunk", link: import.meta.env.VITE_AZURE_2_SPLUNK },
-    { name: "Dynatrace", link: import.meta.env.VITE_DYNATRACE },
-    { name: "Rigor Monitoring", link: import.meta.env.VITE_RIGOR_MONITORING },
-    { name: "Grafana", link: import.meta.env.VITE_GRAFANA },
-    { name: "Alert Manager", link: import.meta.env.VITE_ALERT_MANAGER },
-    { name: "Azure Endeca", link: import.meta.env.VITE_AZURE_EMDECA },
-    { name: "Forter Portal", link: import.meta.env.VITE_FORTER_PORTAL },
-    { name: "BI Rewards", link: import.meta.env.VITE_BI_REWARDS },
-    { name: "ArgoCD", link: import.meta.env.VITE_ARGOCD },
-    { name: "UserAgent Lookup", link: import.meta.env.VITE_USERAGENT_LOOKUP },
-    { name: "PagerDuty", link: import.meta.env.VITE_PAGERDUTY },
-    { name: "Service Now", link: import.meta.env.VITE_SERVICE_NOW },
-    { name: "PAM", link: import.meta.env.VITE_PAM },
-    { name: "Server Restart", link: import.meta.env.VITE_SERVER_RESTART },
-    { name: "SephAdmin", link: import.meta.env.VITE_SEPH_ADMIN },
-    { name: "True Origin", link: import.meta.env.VITE_TRUE_ORIGIN },
-    { name: "True Preview", link: import.meta.env.VITE_TRUE_PREVIEW },
-    { name: "Mule Catalog", link: import.meta.env.VITE_MULE_CATALOG },
-    { name: "Mule BackOffice", link: import.meta.env.VITE_MULE_BACKOFFICE },
-    { name: "BCC", link: import.meta.env.VITE_BCC },
-    { name: "CSC", link: import.meta.env.VITE_CSC },
-  ]
-}
-  , {
-  id: 1,
-  header: "Wiki Links", links: [
-    { name: "Environment URLs", link: import.meta.env.VITE_ENVIRONMENT_URLS },
-    { name: "Support Hub", link: import.meta.env.VITE_SUPPORT_HUB },
-    { name: "Dotcom Incident Report", link: import.meta.env.VITE_DOTCOM_INCIDENT_REPORT },
-    { name: "P1/P2 Checklist", link: import.meta.env.VITE_P1_P2_CHECKLIST },
-  ]
-}
+export const QUICK_LINKS = [
+  {
+    id: 0,
+    header: "Alerting & Monitoring Tools",
+    links: [
+      { name: "Azure Splunk", link: import.meta.env.VITE_AZURE_SPLUNK },
+      { name: "Azure 2 Splunk", link: import.meta.env.VITE_AZURE_2_SPLUNK },
+      { name: "Dynatrace", link: import.meta.env.VITE_DYNATRACE },
+      { name: "Rigor Monitoring", link: import.meta.env.VITE_RIGOR_MONITORING },
+      { name: "Grafana", link: import.meta.env.VITE_GRAFANA },
+      { name: "Alert Manager", link: import.meta.env.VITE_ALERT_MANAGER },
+      { name: "Azure Endeca", link: import.meta.env.VITE_AZURE_EMDECA },
+      { name: "Forter Portal", link: import.meta.env.VITE_FORTER_PORTAL },
+      { name: "BI Rewards", link: import.meta.env.VITE_BI_REWARDS },
+      { name: "ArgoCD", link: import.meta.env.VITE_ARGOCD },
+      { name: "UserAgent Lookup", link: import.meta.env.VITE_USERAGENT_LOOKUP },
+      { name: "PagerDuty", link: import.meta.env.VITE_PAGERDUTY },
+      { name: "Service Now", link: import.meta.env.VITE_SERVICE_NOW },
+      { name: "PAM", link: import.meta.env.VITE_PAM },
+      { name: "Server Restart", link: import.meta.env.VITE_SERVER_RESTART },
+      { name: "SephAdmin", link: import.meta.env.VITE_SEPH_ADMIN },
+      { name: "True Origin", link: import.meta.env.VITE_TRUE_ORIGIN },
+      { name: "True Preview", link: import.meta.env.VITE_TRUE_PREVIEW },
+      { name: "Mule Catalog", link: import.meta.env.VITE_MULE_CATALOG },
+      { name: "Mule BackOffice", link: import.meta.env.VITE_MULE_BACKOFFICE },
+      { name: "BCC", link: import.meta.env.VITE_BCC },
+      { name: "CSC", link: import.meta.env.VITE_CSC },
+    ],
+  },
+  {
+    id: 1,
+    header: "Wiki Links",
+    links: [
+      { name: "Environment URLs", link: import.meta.env.VITE_ENVIRONMENT_URLS },
+      { name: "Support Hub", link: import.meta.env.VITE_SUPPORT_HUB },
+      {
+        name: "Dotcom Incident Report",
+        link: import.meta.env.VITE_DOTCOM_INCIDENT_REPORT,
+      },
+      { name: "P1/P2 Checklist", link: import.meta.env.VITE_P1_P2_CHECKLIST },
+    ],
+  },
 ];
 
-export const QUICK_LINKS_HEADER = "Quick Links"
+export const QUICK_LINKS_HEADER = "Quick Links";
 export const CHART_TABS = [{ header: BAR }, { header: LINE }];
